requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
cattrs==23.2.3
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
distlib==0.3.8
Django==5.0.6
django-allauth==0.63.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-salmonella==1.2.1
django-schema-graph==3.1.0
django-timezone-field==6.1.0
django-unfold==0.29.1
djangorestframework==3.15.2
filelock==3.15.4
flower==2.0.1
humanize==4.10.0
kombu==5.3.7
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pygame==2.5.2
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
